+++ samples/open-telemetry/src/main/java/com/google/cloud/spanner/sample/Application.java
@@ -1,3 +1,0 @@
-package com.google.cloud.spanner.sample;
-
-public class Application {}