+++ src/main/java/com/google/cloud/spanner/jdbc/JdbcResultSet.java
@@ -202,6 +202,7 @@ class JdbcResultSet extends AbstractJdbcResultSet {
       case BOOL:
         return isNull ? null : String.valueOf(spanner.getBoolean(spannerIndex));
       case BYTES:
+      case PROTO:
         return isNull ? null : spanner.getBytes(spannerIndex).toBase64();
       case DATE:
         return isNull ? null : spanner.getDate(spannerIndex).toString();
@@ -210,6 +211,7 @@ class JdbcResultSet extends AbstractJdbcResultSet {
       case FLOAT64:
         return isNull ? null : Double.toString(spanner.getDouble(spannerIndex));
       case INT64:
+      case ENUM:
         return isNull ? null : Long.toString(spanner.getLong(spannerIndex));
       case NUMERIC:
         return isNull ? null : spanner.getBigDecimal(spannerIndex).toString();
@@ -244,6 +246,7 @@ class JdbcResultSet extends AbstractJdbcResultSet {
       case FLOAT64:
         return !isNull && spanner.getDouble(spannerIndex) != 0D;
       case INT64:
+      case ENUM:
         return !isNull && spanner.getLong(spannerIndex) != 0L;
       case NUMERIC:
         return !isNull && !spanner.getBigDecimal(spannerIndex).equals(BigDecimal.ZERO);
@@ -257,6 +260,7 @@ class JdbcResultSet extends AbstractJdbcResultSet {
       case DATE:
       case STRUCT:
       case TIMESTAMP:
+      case PROTO:
       case ARRAY:
       default:
         throw createInvalidToGetAs("boolean", type);
@@ -281,6 +285,7 @@ class JdbcResultSet extends AbstractJdbcResultSet {
             ? (byte) 0
             : checkedCastToByte(Double.valueOf(spanner.getDouble(spannerIndex)).longValue());
       case INT64:
+      case ENUM:
         return isNull ? (byte) 0 : checkedCastToByte(spanner.getLong(spannerIndex));
       case NUMERIC:
         return isNull ? (byte) 0 : checkedCastToByte(spanner.getBigDecimal(spannerIndex));
@@ -296,6 +301,7 @@ class JdbcResultSet extends AbstractJdbcResultSet {
       case DATE:
       case STRUCT:
       case TIMESTAMP:
+      case PROTO:
       case ARRAY:
       default:
         throw createInvalidToGetAs("byte", type);
@@ -320,6 +326,7 @@ class JdbcResultSet extends AbstractJdbcResultSet {
             ? 0
             : checkedCastToShort(Double.valueOf(spanner.getDouble(spannerIndex)).longValue());
       case INT64:
+      case ENUM:
         return isNull ? 0 : checkedCastToShort(spanner.getLong(spannerIndex));
       case NUMERIC:
         return isNull ? 0 : checkedCastToShort(spanner.getBigDecimal(spannerIndex));
@@ -335,6 +342,7 @@ class JdbcResultSet extends AbstractJdbcResultSet {
       case DATE:
       case STRUCT:
       case TIMESTAMP:
+      case PROTO:
       case ARRAY:
       default:
         throw createInvalidToGetAs("short", type);
@@ -359,6 +367,7 @@ class JdbcResultSet extends AbstractJdbcResultSet {
             ? 0
             : checkedCastToInt(Double.valueOf(spanner.getDouble(spannerIndex)).longValue());
       case INT64:
+      case ENUM:
         return isNull ? 0 : checkedCastToInt(spanner.getLong(spannerIndex));
       case NUMERIC:
         return isNull ? 0 : checkedCastToInt(spanner.getBigDecimal(spannerIndex));
@@ -374,6 +383,7 @@ class JdbcResultSet extends AbstractJdbcResultSet {
       case DATE:
       case STRUCT:
       case TIMESTAMP:
+      case PROTO:
       case ARRAY:
       default:
         throw createInvalidToGetAs("int", type);
@@ -394,6 +404,7 @@ class JdbcResultSet extends AbstractJdbcResultSet {
       case FLOAT64:
         return isNull ? 0L : Double.valueOf(spanner.getDouble(spannerIndex)).longValue();
       case INT64:
+      case ENUM:
         return isNull ? 0L : spanner.getLong(spannerIndex);
       case NUMERIC:
         return isNull ? 0 : checkedCastToLong(parseBigDecimal(spanner.getString(spannerIndex)));
@@ -409,6 +420,7 @@ class JdbcResultSet extends AbstractJdbcResultSet {
       case DATE:
       case STRUCT:
       case TIMESTAMP:
+      case PROTO:
       case ARRAY:
       default:
         throw createInvalidToGetAs("long", type);
@@ -429,6 +441,7 @@ class JdbcResultSet extends AbstractJdbcResultSet {
       case FLOAT64:
         return isNull ? 0 : checkedCastToFloat(spanner.getDouble(spannerIndex));
       case INT64:
+      case ENUM:
         return isNull ? 0 : checkedCastToFloat(spanner.getLong(spannerIndex));
       case NUMERIC:
         return isNull ? 0 : spanner.getBigDecimal(spannerIndex).floatValue();
@@ -442,6 +455,7 @@ class JdbcResultSet extends AbstractJdbcResultSet {
       case DATE:
       case STRUCT:
       case TIMESTAMP:
+      case PROTO:
       case ARRAY:
       default:
         throw createInvalidToGetAs("float", type);
@@ -462,6 +476,7 @@ class JdbcResultSet extends AbstractJdbcResultSet {
       case FLOAT64:
         return isNull ? 0 : spanner.getDouble(spannerIndex);
       case INT64:
+      case ENUM:
         return isNull ? 0 : spanner.getLong(spannerIndex);
       case NUMERIC:
         return isNull ? 0 : spanner.getBigDecimal(spannerIndex).doubleValue();
@@ -475,6 +490,7 @@ class JdbcResultSet extends AbstractJdbcResultSet {
       case DATE:
       case STRUCT:
       case TIMESTAMP:
+      case PROTO:
       case ARRAY:
       default:
         throw createInvalidToGetAs("double", type);
@@ -514,6 +530,8 @@ class JdbcResultSet extends AbstractJdbcResultSet {
       case JSON:
       case PG_JSONB:
       case STRUCT:
+      case PROTO:
+      case ENUM:
       case ARRAY:
       default:
         throw createInvalidToGetAs("date", type);
@@ -542,6 +560,8 @@ class JdbcResultSet extends AbstractJdbcResultSet {
       case JSON:
       case PG_JSONB:
       case STRUCT:
+      case PROTO:
+      case ENUM:
       case ARRAY:
       default:
         throw createInvalidToGetAs("time", type);
@@ -571,6 +591,8 @@ class JdbcResultSet extends AbstractJdbcResultSet {
       case JSON:
       case PG_JSONB:
       case STRUCT:
+      case PROTO:
+      case ENUM:
       case ARRAY:
       default:
         throw createInvalidToGetAs("timestamp", type);
@@ -743,6 +765,8 @@ class JdbcResultSet extends AbstractJdbcResultSet {
       return getString(columnIndex);
     }
     if (type == Type.timestamp()) return getTimestamp(columnIndex);
+    if (type.getCode() == Code.PROTO) return getBytes(columnIndex);
+    if (type.getCode() == Code.ENUM) return getLong(columnIndex);
     if (type.getCode() == Code.ARRAY) return getArray(columnIndex);
     throw JdbcSqlExceptionFactory.of("Unknown type: " + type, com.google.rpc.Code.INVALID_ARGUMENT);
   }
@@ -816,6 +840,7 @@ class JdbcResultSet extends AbstractJdbcResultSet {
         res = isNull ? null : BigDecimal.valueOf(spanner.getDouble(spannerIndex));
         break;
       case INT64:
+      case ENUM:
         res = isNull ? null : BigDecimal.valueOf(spanner.getLong(spannerIndex));
         break;
       case NUMERIC:
@@ -840,6 +865,7 @@ class JdbcResultSet extends AbstractJdbcResultSet {
       case DATE:
       case TIMESTAMP:
       case STRUCT:
+      case PROTO:
       case ARRAY:
       default:
         throw createInvalidToGetAs("BigDecimal", type);
@@ -932,6 +958,8 @@ class JdbcResultSet extends AbstractJdbcResultSet {
       case JSON:
       case PG_JSONB:
       case STRUCT:
+      case PROTO:
+      case ENUM:
       case ARRAY:
       default:
         throw createInvalidToGetAs("date", type);
@@ -964,6 +992,8 @@ class JdbcResultSet extends AbstractJdbcResultSet {
       case JSON:
       case PG_JSONB:
       case STRUCT:
+      case PROTO:
+      case ENUM:
       case ARRAY:
       default:
         throw createInvalidToGetAs("time", type);
@@ -999,6 +1029,8 @@ class JdbcResultSet extends AbstractJdbcResultSet {
       case JSON:
       case PG_JSONB:
       case STRUCT:
+      case PROTO:
+      case ENUM:
       case ARRAY:
       default:
         throw createInvalidToGetAs("timestamp", type);